requirements.txt
Django==2.2
django-cors-headers==2.5.2
django-filter==2.1.0
djangorestframework==3.9.2
Pillow==6.0.0
pytz==2018.9
sqlparse==0.3.0
